requirements.txt
flask
flask-sqlalchemy
marshmallow-sqlalchemy
mysqlclient
python-dotenv